# apps/be/static/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDto'
              examples:
                default:
                  value:
                    accessToken: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.abc123
                    refreshToken: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.refresh123
                    tokenType: Bearer
      summary: 로그인
      tags: &ref_0
        - 인증
...
  /api/auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDto'
              examples:
                default:
                  value:
                    accessToken: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.abc123
                    refreshToken: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.refresh123
                    tokenType: Bearer
      summary: 토큰 갱신
      tags: *ref_0
  /api/auth/{provider}:
    post:
      operationId: AuthController_socialLogin
      parameters:
        - name: provider
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDto'
              examples:
                default:
                  value:
                    accessToken: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.abc123
                    refreshToken: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.refresh123
                    tokenType: Bearer
      summary: 소셜 로그인
      tags: *ref_0
  /api/auth/phone/send:
    post:
      operationId: AuthController_sendPhone
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPhoneRequestDto'
      responses:
        '200':
          description: ''
      summary: 휴대전화 인증 요청
      tags: *ref_0
  /api/auth/phone/verify:
    post:
      operationId: AuthController_verifyPhone
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPhoneRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyPhoneResponseDto'
              examples:
                default:
                  value:
                    verificationToken: verify_token_abc123
      summary: 휴대전화 인증 확인
      tags: *ref_0
...
  /api/bakeries/{id}:
    get:
      operationId: BakeriesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BakeryDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    name: 르뺑 베이커리
                    address: 대전시 중구 은행동 123-4
                    lat: 36.328
                    lng: 127.4272
                    thumbnailUrl: https://cdn.breadbread.kr/bakeries/lepain.jpg
                    rating: 4.8
                    reviewCount: 124
                    phone: 042-1234-5678
                    openHours: 09:00 - 21:00
                    description: 은행동의 인기 베이커리입니다.
                    menus:
                      - id: 1
                        name: 크루아상
                        price: 4500
                        imageUrl: https://cdn.breadbread.kr/menus/croissant.jpg
                      - id: 2
                        name: 소금빵
                        price: 3500
                        imageUrl: https://cdn.breadbread.kr/menus/saltbread.jpg
      summary: 빵집 상세 조회
      tags: *ref_2
...
  /api/courses/{id}:
    get:
      operationId: CoursesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    name: 은행동 빵지순례 코스
                    thumbnailUrl: https://cdn.breadbread.kr/courses/eunhaeng.jpg
                    bakeryCount: 4
                    estimatedTime: 120
                    description: 은행동의 숨은 빵집들을 탐방하는 코스입니다.
                    bakeries:
                      - id: 1
                        name: 르뺑 베이커리
                        address: 대전시 중구 은행동 123-4
                        lat: 36.328
                        lng: 127.4272
                        thumbnailUrl: https://cdn.breadbread.kr/bakeries/lepain.jpg
                        rating: 4.8
                        reviewCount: 124
                      - id: 3
                        name: 은행제빵소
                        address: 대전시 중구 은행동 78-2
                        lat: 36.329
                        lng: 127.426
                        thumbnailUrl: https://cdn.breadbread.kr/bakeries/eunhaeng.jpg
                        rating: 4.5
                        reviewCount: 67
      summary: 코스 상세 조회
      tags: *ref_3
  /api/courses/recommend:
    post:
      operationId: CoursesController_recommend
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    name: 은행동 빵지순례 코스
                    thumbnailUrl: https://cdn.breadbread.kr/courses/eunhaeng.jpg
                    bakeryCount: 4
                    estimatedTime: 120
                    description: 은행동의 숨은 빵집들을 탐방하는 코스입니다.
                    bakeries:
                      - id: 1
                        name: 르뺑 베이커리
                        address: 대전시 중구 은행동 123-4
                        lat: 36.328
                        lng: 127.4272
                        thumbnailUrl: https://cdn.breadbread.kr/bakeries/lepain.jpg
                        rating: 4.8
                        reviewCount: 124
                      - id: 3
                        name: 은행제빵소
                        address: 대전시 중구 은행동 78-2
                        lat: 36.329
                        lng: 127.426
                        thumbnailUrl: https://cdn.breadbread.kr/bakeries/eunhaeng.jpg
                        rating: 4.5
                        reviewCount: 67
      summary: AI 코스 추천
      tags: *ref_3
...
  /api/community:
    get:
      operationId: CommunityController_findAll
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponseDto'
              examples:
                default:
                  value:
                    posts:
                      - id: 1
                        title: 은행동 빵집 투어 후기
                        content: 오늘 은행동 빵집 투어를 다녀왔어요. 르뺑 베이커리의 크루아상이 정말 맛있었습니다!
                        imageUrls:
                          - https://cdn.breadbread.kr/community/post1-1.jpg
                        author:
                          id: 1
                          name: 홍길동
                        likeCount: 24
                        commentCount: 5
                        createdAt: '2026-03-27T15:30:00.000Z'
                    total: 1
      summary: 게시글 목록 조회
      tags: &ref_6
        - 빵뮤니티
    post:
      operationId: CommunityController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    title: 은행동 빵집 투어 후기
                    content: 오늘 은행동 빵집 투어를 다녀왔어요. 르뺑 베이커리의 크루아상이 정말 맛있었습니다!
                    imageUrls:
                      - https://cdn.breadbread.kr/community/post1-1.jpg
                    author:
                      id: 1
                      name: 홍길동
                    likeCount: 24
                    commentCount: 5
                    createdAt: '2026-03-27T15:30:00.000Z'
      summary: 게시글 작성
      tags: *ref_6
  /api/community/{id}:
    get:
      operationId: CommunityController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    title: 은행동 빵집 투어 후기
                    content: 오늘 은행동 빵집 투어를 다녀왔어요. 르뺑 베이커리의 크루아상이 정말 맛있었습니다!
                    imageUrls:
                      - https://cdn.breadbread.kr/community/post1-1.jpg
                    author:
                      id: 1
                      name: 홍길동
                    likeCount: 24
                    commentCount: 5
                    createdAt: '2026-03-27T15:30:00.000Z'
      summary: 게시글 상세 조회
      tags: *ref_6
    patch:
      operationId: CommunityController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetailDto'
              examples:
                default:
                  value:
                    id: 1
                    title: 수정된 제목
                    content: 오늘 은행동 빵집 투어를 다녀왔어요. 르뺑 베이커리의 크루아상이 정말 맛있었습니다!
                    imageUrls:
                      - https://cdn.breadbread.kr/community/post1-1.jpg
                    author:
                      id: 1
                      name: 홍길동
                    likeCount: 24
                    commentCount: 5
                    createdAt: '2026-03-27T15:30:00.000Z'
      summary: 게시글 수정
      tags: *ref_6
    delete:
      operationId: CommunityController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: 게시글 삭제
      tags: *ref_6
  /api/community/{id}/comments:
    get:
      operationId: CommunityController_getComments
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponseDto'
              examples:
                default:
                  value:
                    comments:
                      - id: 1
                        content: 저도 다녀왔는데 정말 맛있더라고요!
                        author:
                          id: 2
                          name: 김철수
                        createdAt: '2026-03-27T16:00:00.000Z'
      summary: 댓글 목록 조회
      tags: *ref_6
    post:
      operationId: CommunityController_createComment
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
              examples:
                default:
                  value:
                    id: 1
                    content: 저도 다녀왔는데 정말 맛있더라고요!
                    author:
                      id: 2
                      name: 김철수
                    createdAt: '2026-03-27T16:00:00.000Z'
      summary: 댓글 작성
      tags: *ref_6
...
info:
  title: 빵빵 (BreadBread) API
  description: 빵빵 택시 서비스 API 명세서
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    BakerySummaryDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 르뺑 베이커리
        address:
          type: string
          example: 대전시 중구 은행동 123-4
        lat:
          type: number
          format: double
          example: 36.328
        lng:
          type: number
          format: double
          example: 127.4272
        thumbnailUrl:
          type: string
          example: https://cdn.breadbread.kr/bakeries/lepain.jpg
      required:
        - id
        - name
        - address
        - lat
        - lng
        - thumbnailUrl
...
    LoginRequestDto:
      type: object
      properties:
        loginId:
          type: string
          example: user123
        password:
          type: string
          example: password123!
      required:
        - loginId
        - password
    TokenResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.abc123
        refreshToken:
          type: string
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiaWF0IjoxNzAwMDAwMDAwfQ.refresh123
        tokenType:
          type: string
          default: Bearer
          example: Bearer
      required:
        - accessToken
        - refreshToken
        - tokenType
    RefreshRequestDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.refresh123
      required:
        - refreshToken
    SocialLoginRequestDto:
      type: object
      properties:
        accessToken:
          type: string
          example: ya29.a0AfH6SMBx...
      required:
        - accessToken
    SendPhoneRequestDto:
      type: object
      properties:
        phone:
          type: string
          example: '01012345678'
      required:
        - phone
    VerifyPhoneRequestDto:
      type: object
      properties:
        phone:
          type: string
          example: '01012345678'
        code:
          type: string
          example: '123456'
      required:
        - phone
        - code
    VerifyPhoneResponseDto:
      type: object
      properties:
        verificationToken:
          type: string
          description: 휴대폰 인증 완료 토큰
          example: verify_token_abc123
      required:
        - verificationToken
...
    BakeryMenuDto:
      type: object
      properties:
        name:
          type: string
          example: 크루아상
        price:
          type: number
          example: 4500
        imageUrl:
          type: string
          example: https://cdn.breadbread.kr/menus/croissant.jpg
      required:
        - name
        - price
        - imageUrl
    BakeryDetailDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 르뺑 베이커리
        address:
          type: string
          example: 대전시 중구 은행동 123-4
        lat:
          type: number
          format: double
          example: 36.328
        lng:
          type: number
          format: double
          example: 127.4272
        thumbnailUrl:
          type: string
          example: https://cdn.breadbread.kr/bakeries/lepain.jpg
        description:
          type: string
          example: 은행동 골목에 위치한 프랑스식 베이커리입니다.
        openHours:
          type: string
          example: 08:00 - 21:00
        phone:
          type: string
          example: 02-1234-5678
        menus:
          type: array
          items:
            $ref: '#/components/schemas/BakeryMenuDto'
      required:
        - id
        - name
        - address
        - lat
        - lng
        - thumbnailUrl
        - description
        - openHours
        - phone
        - menus
...
    CourseDetailDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: 은행동 빵지순례 코스
        thumbnailUrl:
          type: string
          example: https://cdn.breadbread.kr/courses/yeonnam.jpg
        bakeryCount:
          type: number
          example: 4
        estimatedTime:
          type: number
          description: 예상 소요 시간 (분)
          example: 120
        description:
          type: string
          example: 은행동 핫플 빵집 4곳을 도는 인기 코스입니다.
        bakeries:
          type: array
          items:
            $ref: '#/components/schemas/BakerySummaryDto'
      required:
        - id
        - name
        - thumbnailUrl
        - bakeryCount
        - estimatedTime
        - description
        - bakeries
    RecommendLocationDto:
      type: object
      properties:
        lat:
          type: number
          format: double
          example: 36.328
        lng:
          type: number
          format: double
          example: 127.4272
      required:
        - lat
        - lng
    RecommendRequestDto:
      type: object
      properties:
        preferences:
          description: '사용자 선호도 (예: 달달한, 건강한, 비건)'
          example:
            - 달달한
            - 촉촉한
          type: array
          items:
            type: string
        maxTime:
          type: number
          description: 최대 이동 시간 (분)
          example: 90
        bakeryCount:
          type: number
          description: 방문할 빵집 수
          example: 3
        location:
          $ref: '#/components/schemas/RecommendLocationDto'
        allergies:
          description: 알러지 정보
          example:
            - 견과류
          type: array
          items:
            type: string
...
    PostSummaryDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        title:
          type: string
          example: 대전 빵집 투어 후기
        authorName:
          type: string
          example: 홍길동
        likeCount:
          type: number
          example: 24
        commentCount:
          type: number
          example: 5
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00.000Z'
      required:
        - id
        - title
        - authorName
        - likeCount
        - commentCount
        - createdAt
    PostListResponseDto:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: '#/components/schemas/PostSummaryDto'
        total:
          type: number
          example: 128
      required:
        - posts
        - total
    CreatePostRequestDto:
      type: object
      properties:
        title:
          type: string
          example: 대전 빵집 투어 후기
        content:
          type: string
          example: 오늘 대전 빵지순례를 다녀왔습니다. 성심당, 메이브레드, 라비에벨을 들렀는데...
        imageUrls:
          example:
            - https://cdn.breadbread.kr/posts/img1.jpg
          type: array
          items:
            type: string
      required:
        - title
        - content
    PostDetailDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        title:
          type: string
          example: 대전 빵집 투어 후기
        authorName:
          type: string
          example: 홍길동
        likeCount:
          type: number
          example: 24
        commentCount:
          type: number
          example: 5
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00.000Z'
        content:
          type: string
          example: 오늘 대전 빵지순례를 다녀왔습니다. 성심당, 메이브레드, 라비에벨을 들렀는데...
        imageUrls:
          example:
            - https://cdn.breadbread.kr/posts/img1.jpg
          type: array
          items:
            type: string
        liked:
          type: boolean
          example: false
      required:
        - id
        - title
        - authorName
        - likeCount
        - commentCount
        - createdAt
        - content
        - imageUrls
        - liked
    UpdatePostRequestDto:
      type: object
      properties:
        title:
          type: string
          example: 대전 빵집 투어 후기 (수정)
        content:
          type: string
          example: 수정된 내용입니다.
        imageUrls:
          example:
            - https://cdn.breadbread.kr/posts/img2.jpg
          type: array
          items:
            type: string
    CommentDto:
      type: object
      properties:
        id:
          type: number
          example: 1
        authorName:
          type: string
          example: 김철수
        content:
          type: string
          example: 저도 다녀왔는데 정말 맛있었어요!
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T11:00:00.000Z'
      required:
        - id
        - authorName
        - content
        - createdAt
    CommentListResponseDto:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentDto'
      required:
        - comments
    CreateCommentRequestDto:
      type: object
      properties:
        content:
          type: string
          example: 저도 다녀왔는데 정말 맛있었어요!
      required:
        - content
...